+++ lib/src/gretl_func.h
@@ -110,7 +110,7 @@ int write_function_package (fnpkg *pkg,
 			    const char *date,
 			    const char *descrip,
 			    FuncDataReq dreq,
-			    float minver);
+			    int minver);
 
 int function_package_get_info (const char *fname,
 			       fnpkg **ppkg,
@@ -121,10 +121,10 @@ int function_package_get_info (const char *fname,
 			       char **date,
 			       char **descrip,
 			       FuncDataReq *dreq,
-			       float *minver);
+			       int *minver);
 
 int check_function_needs (const DATAINFO *pdinfo, FuncDataReq dreq,
-			  float minver);
+			  int minver);
 
 int write_user_function_file (const char *fname);
